refactor: remove db init to a top level persistence file
