+++ src/monitor/persistence.rs
@@ -1,6 +1,4 @@
-use sqlx::sqlite::{SqliteConnectOptions, SqliteJournalMode};
-use sqlx::{ConnectOptions, SqliteConnection};
-use std::str::FromStr;
+use sqlx::SqliteConnection;
 
 mod monitor_config;
 use self::monitor_config::create_monitor_configs_table;
@@ -30,53 +28,23 @@ mod status_mem;
 use self::status_mem::{create_mem_status_frame_singles_table, create_mem_status_frames_table};
 pub use self::status_mem::{get_mem_status_between_dates, insert_mem_status_frame};
 
-const SQLITE_DBS_FOLDER_PATH: &str = "./db";
-const SQLITE_DB_PATH: &str = "./db/monitor.sqlite3";
-const SQLITE_DB_CONN_STR: &str = "sqlite:./db/monitor.sqlite3";
+pub use crate::persistence::{get_default_sql_connection, get_sql_connection, FetchId};
 
-#[derive(Debug, sqlx::FromRow)]
-struct FetchId {
-    pub id: i64,
-}
+pub async fn init_db(conn: &mut SqliteConnection) -> Result<(), sqlx::Error> {
+    create_monitor_configs_table(conn).await?;
 
-async fn get_sql_connection(db_path: &str) -> Result<SqliteConnection, sqlx::Error> {
-    let conn = SqliteConnectOptions::from_str(db_path)
-        .unwrap()
-        .journal_mode(SqliteJournalMode::Wal)
-        .connect()
-        .await?;
+    create_hardware_cpu_infos_table(conn).await?;
+    create_hardware_disk_infos_table(conn).await?;
+    create_hardware_mem_infos_table(conn).await?;
 
-    Ok(conn)
-}
+    create_cpu_status_frames_table(conn).await?;
+    create_cpu_status_frame_cores_table(conn).await?;
 
-pub async fn init_db() -> Result<(), sqlx::Error> {
-    // check if db folder exists
-    if !std::path::Path::new(SQLITE_DBS_FOLDER_PATH).exists() {
-        // create db folder
-        std::fs::create_dir(SQLITE_DBS_FOLDER_PATH).unwrap();
-    }
-    // if db file not exists, create it
-    if !std::path::Path::new(SQLITE_DB_PATH).exists() {
-        // create db file
-        std::fs::File::create(SQLITE_DB_PATH).unwrap();
-    }
+    create_disk_status_frames_table(conn).await?;
+    create_disk_status_frame_singles_table(conn).await?;
 
-    let mut conn = get_sql_connection(SQLITE_DB_CONN_STR).await?;
-
-    create_monitor_configs_table(&mut conn).await?;
-
-    create_hardware_cpu_infos_table(&mut conn).await?;
-    create_hardware_disk_infos_table(&mut conn).await?;
-    create_hardware_mem_infos_table(&mut conn).await?;
-
-    create_cpu_status_frames_table(&mut conn).await?;
-    create_cpu_status_frame_cores_table(&mut conn).await?;
-
-    create_disk_status_frames_table(&mut conn).await?;
-    create_disk_status_frame_singles_table(&mut conn).await?;
-
-    create_mem_status_frames_table(&mut conn).await?;
-    create_mem_status_frame_singles_table(&mut conn).await?;
+    create_mem_status_frames_table(conn).await?;
+    create_mem_status_frame_singles_table(conn).await?;
 
     Ok(())
 }
