+++ src/monitor/persistence/hardware_cpu_info.rs
@@ -2,12 +2,12 @@ use sqlx::SqliteConnection;
 
 use crate::monitor::models::get_hardware_info::HardwareCpuInfo;
 
-use super::{get_sql_connection, FetchId, SQLITE_DB_CONN_STR};
+use super::{get_default_sql_connection, FetchId};
 
 const HARDWARE_CPU_INFOS_TABLE_NAME: &str = "cpu_infos";
 
 pub(super) async fn insert_hardware_cpu_info(info: &HardwareCpuInfo) -> Result<(), sqlx::Error> {
-    let mut conn = get_sql_connection(SQLITE_DB_CONN_STR).await?;
+    let mut conn = get_default_sql_connection().await?;
 
     // check if a record with the same cpu_id already exists
     let exists_record_check = format!(
@@ -58,7 +58,7 @@ pub(super) async fn insert_hardware_cpu_info(info: &HardwareCpuInfo) -> Result<(
 }
 
 pub(super) async fn fetch_latest_hardware_cpus_info() -> Result<Vec<HardwareCpuInfo>, sqlx::Error> {
-    let mut conn = get_sql_connection(SQLITE_DB_CONN_STR).await?;
+    let mut conn = get_default_sql_connection().await?;
 
     let statement = format!(
         "
